+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bug-fixes
 
+- [x] [#121 - Popup open to large on big screens](https://github.com/scriptPilot/app-framework/issues/121)
 - [x] [#557 - Dark brackground for application limits lost](https://github.com/scriptPilot/app-framework/issues/557)
 - [x] [#560 - Error on "npm run android"](https://github.com/scriptPilot/app-framework/issues/560)
 - [x] [#566 - App cache does not work in online mode](https://github.com/scriptPilot/app-framework/issues/566)